requirements.txt
streamlit 
ultralytics 
pillow 
numpy 
